using useQuery in role page
