+++ package.json
@@ -28,6 +28,7 @@
     "dotenv": "^16.4.7",
     "jose": "^6.0.10",
     "js-cookie": "^3.0.5",
+    "js-cookies": "^1.0.4",
     "lucide-react": "^0.484.0",
     "mongodb": "^6.15.0",
     "mongoose": "^8.13.1",
