+++ src/db_script/update.js
@@ -1,0 +1,10 @@
+import User from "../schema/user.ts"
+import connectDb from "../lib/mongoDb.js"
+
+await connectDb()
+
+
+User.updateMany(
+    { role: { $exists: false } },
+    { $set: { role: "Free Plan" } },
+)
