+++ src/schema/user.ts
@@ -10,7 +10,7 @@ const userSchema = new mongoose.Schema({
     createdAt : {type : Date , default : Date.now},
     location : { type : String, default: ""},
     avatarUrl : { type : String, default: ""},
-    role : { type : String, default: "normal"},
+    role : { type : String, default: "Free Plan"},
     bio : { type : String, default: ""},
     backgroundUrl : { type : String, default: ""},
 },{
